Create routes & move main logics
