+++ src/pages/New.jsx
@@ -1,0 +1,12 @@
+import { Link } from 'react-router-dom'
+
+const New = () => {
+  return (
+    <>
+      <Link to="/">Back to Main</Link>
+      <div>New Todo form</div>
+    </>
+  )
+}
+
+export default New
